Updated the map slice to handle filters
