+++ src/UI/components/map/utils/map.utils.test.ts
@@ -85,64 +85,3 @@ describe(responseToGEOJSON.name, () => {
     expect(testResponseToGEOJSON).toEqual(expectedGEOJSON);
   });
 });
-
-import { unpackOverlays } from './map.utils';
-
-describe(unpackOverlays.name, () => {
-  it('produces an array of additional overlays', () => {
-    const test_overlays = [
-      {
-        name: 'an_gambiae',
-        sourceLayer: 'overlays',
-        sourceType: 'raster',
-      },
-      {
-        name: 'world',
-        sourceType: 'vector',
-        overlays: [
-          { name: 'countries' },
-          { name: 'land' },
-          { name: 'oceans' },
-          { name: 'rivers_lakes' },
-        ],
-      },
-    ];
-    const overlays = unpackOverlays(test_overlays).find(
-      (l: any) => l.sourceLayer === 'overlays'
-    );
-    const numOverlays = test_overlays.filter(
-      (l: any) => l.sourceLayer === 'overlays'
-    ).length;
-    expect(overlays.length === numOverlays);
-  });
-  it('produces an array of base map overlays', () => {
-    const test_overlays = [
-      {
-        name: 'an_gambiae',
-        sourceLayer: 'overlays',
-        sourceType: 'raster',
-      },
-      {
-        name: 'world',
-        sourceType: 'vector',
-        overlays: [
-          { name: 'countries' },
-          { name: 'land' },
-          { name: 'oceans' },
-          { name: 'rivers_lakes' },
-        ],
-      },
-    ];
-    const overlays = unpackOverlays(test_overlays).find(
-      (l: any) => l.sourceLayer === 'world'
-    );
-    const numOverlays = test_overlays.filter(
-      (l: any) => l.sourceLayer === 'world'
-    ).length;
-    expect(overlays.length === numOverlays);
-  });
-  it('handles empty argument correctly', () => {
-    const unpackedOverlays = unpackOverlays([]);
-    expect(unpackedOverlays !== (null || undefined));
-  });
-});
