+++ src/UI/package.json
@@ -21,9 +21,11 @@
     "@hookform/resolvers": "^2.9.8",
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.9.3",
+    "@mui/x-date-pickers": "^5.0.6",
     "@reduxjs/toolkit": "^1.8.3",
     "@typescript-eslint/eslint-plugin": "^5.30.7",
     "axios": "^0.27.2",
+    "date-fns": "^2.29.3",
     "js-file-download": "^0.4.12",
     "leaflet": "^1.8.0",
     "mapbox-gl": "^2.9.1",
