+++ src/UI/state/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import configReducer from './configSlice';
-import mapReducer from './mapSlice';
+import mapReducer from './map/mapSlice';
 import authReducer from './authSlice';
 import sourceReducer from './sourceSlice';
